Global names refactoring

+++ kernel/aux/argtable/argtable.c
@@ -28,7 +28,8 @@ USA.
 #include <string.h>
 #include <ctype.h>
 #include <errno.h>
-#include "common/argtable.h"
+
+#include "argtable.h"
 
 #define ARGSTRLEN 1024
 #define NAMESTRING(A) ((A.argname)?(A.argname):(arg_typestr[A.argtype]))
@@ -839,7 +840,7 @@ int (arg_scanargv)(int argc,          /**< number of entries in 'argv'. */
   if (*p!='\0')
      {
      if (ErrMsg)
-         sprintf(ErrMsg,"unexpected argument: %.*s", strcspn(p, " \n\t"), p);
+         sprintf(ErrMsg,"unexpected argument: %.*s", (int) strcspn(p, " \n\t"), p);
      if (ErrMark)
         {
         size_t n = strcspn(p,whitespace);
